refactor(primitives): combine blob and l1 cost helpers
