+++ crates/extractor/src/lib.rs
@@ -14,17 +14,13 @@ use alloy::{
     primitives::{Address, BlockHash, BlockNumber},
     providers::{Provider, ProviderBuilder},
 };
-use alloy_consensus::{BlockHeader, Transaction};
-use alloy_network_primitives::ReceiptResponse;
+use alloy_consensus::BlockHeader;
 use alloy_rpc_client::ClientBuilder;
 use chainio::TaikoInbox;
 use derive_more::Debug;
 use eyre::Result;
 use network::retries::{DEFAULT_RETRY_LAYER, RetryWsConnect};
-use primitives::{
-    blob_fee::calculate_blob_fee_from_receipt,
-    headers::{L1Header, L1HeaderStream, L2Header, L2HeaderStream},
-};
+use primitives::headers::{L1Header, L1HeaderStream, L2Header, L2HeaderStream};
 use std::time::Duration;
 use tokio::{sync::mpsc, time::sleep};
 use tokio_stream::{Stream, StreamExt, wrappers::UnboundedReceiverStream};
@@ -419,6 +415,7 @@ impl Extractor {
         inbox: Address,
     ) -> Result<u128> {
         use alloy_rpc_types_eth::BlockId;
+        use primitives::l1_data_cost::compute_l1_data_posting_cost;
 
         let block = self
             .l1_provider
@@ -429,23 +426,8 @@ impl Extractor {
         let receipts_opt = self.l1_provider.get_block_receipts(BlockId::hash(block_hash)).await?;
         let receipts = receipts_opt.ok_or_else(|| eyre::eyre!("missing receipts"))?;
 
-        let mut total = 0u128;
         let txs = block.transactions.into_transactions_vec();
-        for (tx, receipt) in txs.into_iter().zip(receipts.into_iter()) {
-            if let Some(h) = tx.blob_versioned_hashes() {
-                if !h.is_empty() && tx.to() == Some(inbox) {
-                    // Regular gas cost
-                    let gas_cost =
-                        (receipt.gas_used() as u128).saturating_mul(receipt.effective_gas_price());
-
-                    // Blob data fee calculation
-                    let blob_data_fee = calculate_blob_fee_from_receipt(&receipt);
-
-                    total = total.saturating_add(gas_cost).saturating_add(blob_data_fee);
-                }
-            }
-        }
-        Ok(total)
+        Ok(compute_l1_data_posting_cost(&txs, &receipts, inbox))
     }
 }
