+++ crates/primitives/src/lib.rs
@@ -1,9 +1,9 @@
 //! Core primitives for the Taikoscope project.
-/// Blob fee calculations
-pub mod blob_fee;
 /// Block analytics helpers
 pub mod block_stats;
 /// Hardware cost estimates
 pub mod hardware;
 /// Block header types
 pub mod headers;
+/// L1 data cost calculation helpers
+pub mod l1_data_cost;
